Coefficient of Quartile Variation (cqv)

Versatile function for the coefficient of quartile variation (cqv)

Arguments

x

An R object. Currently there are methods for numeric vectors

na.rm

a logical value indicating whether NA values should be stripped before the computation proceeds.

digits

integer indicating the number of decimal places to be used.

method

a scalar representing the type of confidence intervals required. The value should be any of the values "bonett", "norm", "basic", "perc", "bca" or "all".

R

integer indicating the number of bootstrap replicates.

Value

An object of type "list" which contains the estimate, the intervals, and the computation method. It has two components:

$method

A description of statistical method used for the computations.

$statistics

A data frame representing three vectors: est, lower and upper limits of 95% confidence interval (CI):

est: cqv*100

Bonett 95% CI: It uses a centering adjustment which helps to equalize the tail error probabilities [1, 2].

Normal approximation 95% CI: The intervals calculated by the normal approximation [3, 4], using boot.ci.

Basic bootstrap 95% CI: The intervals calculated by the basic bootstrap method [3, 4], using boot.ci.

Bootstrap percentile 95% CI: The intervals calculated by the bootstrap percentile method [3, 4], using boot.ci.

Adjusted bootstrap percentile (BCa) 95% CI: The intervals calculated by the adjusted bootstrap percentile (BCa) method [3, 4], using boot.ci.

Details

Coefficient of Quartile Variation

The cqv is a measure of relative dispersion that is based on interquartile range (iqr). Since \(cqv\) is unitless, it is useful for comparison of variables with different units. It is also a measure of homogeneity [1, 2].

References

[1] Bonett, DG., 2006, Confidence interval for a coefficient of quartile variation, Computational Statistics & Data Analysis, 50(11), 2953-7, DOI: doi:10.1016/j.csda.2005.05.007

[2] Altunkaynak, B., Gamgam, H., 2018, Bootstrap confidence intervals for the coefficient of quartile variation, Simulation and Computation, 1-9, DOI: doi:10.1080/03610918.2018.1435800

[3] Canty, A., & Ripley, B, 2017, boot: Bootstrap R (S-Plus) Functions. R package version 1.3-20.

[4] Davison, AC., & Hinkley, DV., 1997, Bootstrap Methods and Their Applications. Cambridge University Press, Cambridge. ISBN 0-521-57391-2

Examples

x <- c(
    0.2, 0.5, 1.1, 1.4, 1.8, 2.3, 2.5, 2.7, 3.5, 4.4,
    4.6, 5.4, 5.4, 5.7, 5.8, 5.9, 6.0, 6.6, 7.1, 7.9
)
cqv_versatile(x)
#> $method
#> [1] "cqv = (q3-q1)/(q3+q1)"
#> 
#> $statistics
#>    est
#>   45.6
#> 
cqv_versatile(x, na.rm = TRUE, digits = 2)
#> $method
#> [1] "cqv = (q3-q1)/(q3+q1)"
#> 
#> $statistics
#>     est
#>   45.62
#> 
cqv_versatile(x, na.rm = TRUE, digits = 2, method = "bonett")
#> $method
#> [1] "cqv with Bonett 95% CI"
#> 
#> $statistics
#>     est lower upper
#>   45.62 24.78 77.33
#>