Coefficient of Variation (cv)

Versatile function for the coefficient of variation (cv)

Arguments

x

An R object. Currently there are methods for numeric vectors

na.rm

a logical value indicating whether NA values should be stripped before the computation proceeds.

digits

integer indicating the number of decimal places to be used.

method

a scalar representing the type of confidence intervals required. The value should be any of the values "kelley", "mckay", "miller", "vangel", "mahmoudvand_hassani", "equal_tailed", "shortest_length", "normal_approximation", "norm","basic", "aak_adj", "aak_ls", "aak_als", or "all".

correction

returns the unbiased estimate of the coefficient of variation

alpha

The allowed type I error probability

R

integer indicating the number of bootstrap replicates.

Value

An object of type "list" which contains the estimate, the intervals, and the computation method. It has two main components:

$method

A description of statistical method used for the computations.

$statistics

A data frame representing three vectors: est, lower and upper limits of confidence interval (CI); additional description vector is provided when "all" is selected:

est: cv*100

Kelley Confidence Interval: Thanks to package MBESS [2] for the computation of confidence limits for the noncentrality parameter from a t distribution conf.limits.nct [3].

McKay Confidence Interval: The intervals calculated by the method introduced by McKay [4], using chi-square distribution.

Miller Confidence Interval: The intervals calculated by the method introduced by Miller [5], using the standard normal distribution.

Vangel Confidence Interval: Vangel [6] proposed a method for the calculation of CI for cv; which is a modification on McKay’s CI.

Mahmoudvand-Hassani Confidence Interval: Mahmoudvand and Hassani [7] proposed a new CI for cv; which is obtained using ranked set sampling (RSS)

Normal Approximation Confidence Interval: Wararit Panichkitkosolkul [8] proposed another CI for cv; which is a normal approximation.

Shortest-Length Confidence Interval: Wararit Panichkitkosolkul [8] proposed another CI for cv; which is obtained through minimizing the length of CI.

Equal-Tailed Confidence Interval: Wararit Panichkitkosolkul [8] proposed another CI for cv; which is obtained using chi-square distribution.

Bootstrap Confidence Intervals: Thanks to package boot by Canty & Ripley [9] we can obtain bootstrap CI around cv using boot.ci.

Abu-Shawiesh-Akyuz-Kibria Confidence Intervals: Abu-Shawiesh, Akyuz, & Kibria [10] proposed three CIs for the population cv that adjust for non-normality through the sample kurtosis: the adjusted-degrees-of-freedom CI (aak_adj), the large-sample CI (aak_ls), and the augmented-large-sample CI (aak_als).

Details

Coefficient of Variation

The cv is a measure of relative dispersion representing the degree of variability relative to the mean [1]. Since \(cv\) is unitless, it is useful for comparison of variables with different units. It is also a measure of homogeneity [1].

References

[1] Albatineh, AN., Kibria, BM., Wilcox, ML., & Zogheib, B, 2014, Confidence interval estimation for the population coefficient of variation using ranked set sampling: A simulation study, Journal of Applied Statistics, 41(4), 733–751, DOI: doi:10.1080/02664763.2013.847405

[2] Kelley, K., 2018, MBESS: The MBESS R Package. R package version 4.4. 3.

[3] Kelley, K., 2007, Sample size planning for the coefficient of variation from the accuracy in parameter estimation approach, Behavior Research Methods, 39(4), 755–766, DOI: doi:10.3758/BF03192966

[4] McKay, AT., 1932, Distribution of the Coefficient of Variation and the Extended“ t” Distribution, Journal of the Royal Statistical Society, 95(4), 695–698

[5] Miller, E., 1991, Asymptotic test statistics for coefficients of variation, Communications in Statistics-Theory and Methods, 20(10), 3351–3363

[6] Vangel, MG., 1996, Confidence intervals for a normal coefficient of variation, The American Statistician, 50(1), 21–26

[7] Mahmoudvand, R., & Hassani, H., 2009, Two new confidence intervals for the coefficient of variation in a normal distribution, Journal of Applied Statistics, 36(4), 429–442

[8] Panichkitkosolkul, W., 2013, Confidence Intervals for the Coefficient of Variation in a Normal Distribution with a Known Population Mean, Journal of Probability and Statistics, 2013, 1–11, doi:10.1155/2013/324940

[9] Canty, A., & Ripley, B., 2017, boot: Bootstrap R (S-Plus) Functions, R package version 1.3-20

[10] Abu-Shawiesh, MOA., Akyuz, HE., & Kibria, BG., 2019, Performance of Some Confidence Intervals for Estimating the Population Coefficient of Variation under both Symmetric and Skewed Distributions, Statistics, Optimization & Information Computing, 7(2), 277-290, DOI: doi:10.19139/soic.v7i2.630

Examples

x <- c(
    0.2, 0.5, 1.1, 1.4, 1.8, 2.3, 2.5, 2.7, 3.5, 4.4,
    4.6, 5.4, 5.4, 5.7, 5.8, 5.9, 6.0, 6.6, 7.1, 7.9
)
cv_versatile(x)
#> $method
#> [1] "cv = sd/mean (may be biased)"
#> 
#> $statistics
#>    est
#>   57.8
#> 
cv_versatile(x, correction = TRUE)
#> $method
#> [1] "Corrected (i.e., unbiased) cv"
#> 
#> $statistics
#>    est
#>   58.1
#> 
cv_versatile(x, na.rm = TRUE, digits = 3, method = "kelley", correction = TRUE)
#> $method
#> [1] "Corrected cv with Kelley 95% CI"
#> 
#> $statistics
#>      est  lower upper
#>   58.058 41.467 98.51
#> 
cv_versatile(x, na.rm = TRUE, method = "mahmoudvand_hassani", correction = TRUE)
#> $method
#> [1] "Corrected cv with Mahmoudvand-Hassani 95% CI"
#> 
#> $statistics
#>    est lower upper
#>   58.1  43.7  83.3
#>